+++ extensions/python/src/language_servers.rs
@@ -1,4 +1,5 @@
 mod basedpyright;
+mod commons;
 mod pyright;
 
 pub use basedpyright::*;
