+++ extensions/python/src/language_servers/pyright.rs
@@ -1,9 +1,15 @@
-use zed::lsp::{Completion, CompletionKind, Symbol, SymbolKind};
-use zed::{CodeLabel, CodeLabelSpan};
-use zed_extension_api::settings::LspSettings;
+use crate::language_servers::commons::CommonPythonLSP;
+use zed::lsp::{Completion, Symbol};
+use zed::CodeLabel;
 use zed_extension_api::{self as zed, serde_json, Result};
 
-pub struct Pyright {}
+pub struct Pyright;
+
+impl CommonPythonLSP for Pyright {
+    fn get_language_server_id(&self) -> &'static str {
+        Self::LANGUAGE_SERVER_ID
+    }
+}
 
 impl Pyright {
     pub const LANGUAGE_SERVER_ID: &'static str = "pyright";
@@ -20,87 +26,18 @@ impl Pyright {
         Ok(path)
     }
 
-    pub fn label_for_completion(&self, completion: Completion) -> Option<CodeLabel> {
-        let highlight_name = match completion.kind? {
-            CompletionKind::Class => "type",
-            CompletionKind::Constant => "constant",
-            CompletionKind::Method => "function.method",
-            CompletionKind::Function => "function",
-            _ => return None,
-        };
-
-        let len = completion.label.len();
-        let name_span = CodeLabelSpan::literal(completion.label, Some(highlight_name.to_string()));
-
-        Some(CodeLabel {
-            code: Default::default(),
-            spans: if let Some(detail) = completion.detail {
-                vec![
-                    name_span,
-                    CodeLabelSpan::literal(" ", None),
-                    CodeLabelSpan::literal(detail, None),
-                ]
-            } else {
-                vec![name_span]
-            },
-            filter_range: (0..len).into(),
-        })
-    }
-
-    pub fn label_for_symbol(&self, symbol: Symbol) -> Option<CodeLabel> {
-        let name = &symbol.name;
-
-        return match symbol.kind {
-            SymbolKind::Method | SymbolKind::Function => {
-                let def = "def ";
-                let code = format!("{def}{name}():\n");
-                let filter_range = 0..name.len();
-                let display_range = def.len()..def.len() + name.len();
-
-                Some(CodeLabel {
-                    code,
-                    spans: vec![CodeLabelSpan::code_range(display_range)],
-                    filter_range: filter_range.into(),
-                })
-            }
-            SymbolKind::Class => {
-                let class = "class ";
-                let code = format!("{class}{name}");
-                let filter_range = 0..name.len();
-                let display_range = class.len()..class.len() + name.len();
-
-                Some(CodeLabel {
-                    code,
-                    spans: vec![CodeLabelSpan::code_range(display_range)],
-                    filter_range: filter_range.into(),
-                })
-            }
-            SymbolKind::Constant => {
-                let code = name.to_uppercase().to_string();
-                let filter_range = 0..name.len();
-                let display_range = 0..name.len();
-
-                Some(CodeLabel {
-                    code,
-                    spans: vec![CodeLabelSpan::code_range(display_range)],
-                    filter_range: filter_range.into(),
-                })
-            }
-            _ => None,
-        };
-    }
-
     pub fn workspace_configuration(
         &mut self,
         worktree: &zed::Worktree,
     ) -> Result<Option<serde_json::Value>> {
-        let settings = LspSettings::for_worktree(Self::LANGUAGE_SERVER_ID, worktree)
-            .ok()
-            .and_then(|lsp_settings| lsp_settings.settings.clone())
-            .unwrap_or_default();
+        CommonPythonLSP::workspace_configuration(self, worktree)
+    }
 
-        Ok(Some(serde_json::json!({
-            Self::LANGUAGE_SERVER_ID: settings
-        })))
+    pub fn label_for_completion(&self, completion: Completion) -> Option<CodeLabel> {
+        CommonPythonLSP::label_for_completion(self, completion)
+    }
+
+    pub fn label_for_symbol(&self, symbol: Symbol) -> Option<CodeLabel> {
+        CommonPythonLSP::label_for_symbol(self, symbol)
     }
 }
